+++ CHANGELOG.md
@@ -71,6 +71,7 @@ CHANGELOG
   - Use a requests session (#1363).
   - Set the requests parameters once per session.
 - `intelmq.bots.experts.maxmind_geoip.expert`: New parameter `use_registered` to use the registered country (#1344).
+- `intelmq.bots.experts.filter.expert`: Support for paths (#1208).
 
 #### Outputs
 - added `intelmq.bots.experts.mcafee.output_esm` (1265).